trying to make

+++ src/main/java/ru/kata/spring/boot_security/demo/model/User.java
@@ -14,7 +14,7 @@ public class User implements UserDetails {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "firstName")
     private String firstName;
@@ -57,11 +57,11 @@ public class User implements UserDetails {
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
